Attempt to deal with Error Handling + prints
When a task finds an error it raises it immediately in the calling
function, but DON'T stop the other tasks, which I assume will still
return their values. That would mean that the total return of gather
are only coroutines that succeded in running, which is good.
I'm assuming I can catch those errors and raise (and deal with) them
in their caller functions allowing the program to continue.